requirements.txt
asgiref==3.5.0
calendra==6.1.2
certifi==2021.10.8
Django==4.0.3
django-multiselectfield==0.1.12
djangorestframework==3.13.1
drf-writable-nested==0.6.3
fiscalyear==0.4.0
jplephem==2.17
jsonfield==3.1.0
lunardate==0.2.0
more-itertools==8.12.0
numpy==1.22.3
pyCalverter==1.6.1
pyluach==1.4.1
python-dateutil==2.8.2
pytz==2022.1
sgp4==2.21
six==1.16.0
skyfield==1.42
skyfield-data==4.0.0
sqlparse==0.4.2
tzdata==2022.1
